Instructions for missings vizualisations

Instructions for missings vizualisations

Usage

missings_specification(
  vline = FALSE,
  hline = FALSE,
  values = NULL,
  drop_from_plot = FALSE
)

Arguments

vline

(logical) Should the vertical line be shown? Defaults to FALSE.

hline

(logical) Should the horizontal line be shown? Defaults to FALSE.

values

(NULL) or single value or a named vector. Specifies the value(-s) that stand for the missing values. If NULL, then no missing value handling is carried out. If single value, then it is assumed that this value is used for flagging missing values across all continuous variables. If named vector, then the names are used to refer to continuous variables and the values for flagging missing values in that variable.

drop_from_plot

(logical) Should the missing values be dropped from plot? Defaults to FALSE.

Value

List of class missings_specification.

Examples

if (FALSE) { # \dontrun{
g_ice(
  sculpture,
  missings_spec = missings_specification(
    vline = TRUE, # show vertical line
    values = -1 # NAs in all continuous variables displayed as -1
  )
)
} # }